requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=6.0

# Chemistry libraries
rdkit-pypi>=2022.0.0
meeko>=0.5.0
biopython>=1.79

# Data processing
scipy>=1.7.0

# Optional: for advanced features
# openbabel-wheel>=0.1.0  # If needed for molecule conversion










